+++ .travis.yml
@@ -1,13 +1,24 @@
+sudo: false
 language: node_js
 node_js:
-  - "4.0"
+  - 4
+  - 5
 
-branches:
-  only:
-    - master
-
+# Make sure we have new NPM.
 before_install:
-  - npm i -g npm
-  # Workaround for a permissions issue with Travis virtual machine images
+  - npm install -g npm
+
 script:
+  - npm run lint
   - npm test
+  - npm run coverage
+
+addons:
+  firefox: 'latest'
+
+before_script:
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
+
+after_success:
+  - npm run coverage-publish
