+++ package.json
@@ -2,9 +2,16 @@
   "name": "libp2p-record",
   "version": "0.1.2",
   "description": "IPRS Record implementation",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/lab tests/*test.js"
+    "test": "aegir-test",
+    "lint": "aegir-lint",
+    "test:node": "aegir-test node",
+    "test:browser": "aegir-test browser",
+    "build": "aegir-build",
+    "release": "aegir-release",
+    "coverage": "aegir-coverage",
+    "coverage-publish": "aegir-coverage publish"
   },
   "repository": {
     "type": "git",
@@ -16,7 +23,10 @@
   "engines": {
     "node": "^4.0.0"
   },
-  "pre-commit": [],
+  "pre-commit": [
+    "lint",
+    "test"
+  ],
   "author": "David Dias <daviddias@ipfs.io>",
   "license": "MIT",
   "bugs": {
@@ -24,15 +34,15 @@
   },
   "homepage": "https://github.com/diasdavid/js-libp2p-record",
   "devDependencies": {
-    "code": "^1.5.0",
-    "lab": "^5.15.2",
-    "pre-commit": "^1.1.1",
-    "standard": "^5.1.1",
-    "merkledag-store": "^0.1.0"
+    "aegir": "^2.1.1",
+    "chai": "^3.5.0",
+    "merkledag-store": "^0.1.0",
+    "pre-commit": "^1.1.2"
   },
   "dependencies": {
     "multihashing": "^0.1.3",
     "ecdsa": "^0.6.0",
     "ipld": "^0.4.0"
-  }
+  },
+  "jsnext:main": "src/index.js"
 }
